+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for resp
 
+## Unreleased
+
+* Change name of type RespReply to RespMessage
+* Change name of RespExpr data constructor to RespReply
+
 ## 1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.